--- src/main/java/com/example/demo/model/Branch.java
@@ -1,7 +1,0 @@
-package com.example.demo.model;
-
-public enum Branch {
-    LOS_ANGELES1,
-    LOS_ANGELES2,
-    LOS_ANGELES3
-}
