--- src/main/java/com/example/demo/model/FuelType.java
@@ -1,7 +1,0 @@
-package com.example.demo.model;
-
-public enum FuelType {
-    GASOLINE,
-    DIESEL,
-    ELECTRIC
-}
