--- src/main/java/com/example/demo/model/ReservationStatus.java
@@ -1,7 +1,0 @@
-package com.example.demo.model;
-
-public enum ReservationStatus {
-    RESERVED,
-    PICKED_UP,
-    COMPLETED
-}
