--- src/main/java/com/example/demo/model/Role.java
@@ -1,6 +1,0 @@
-package com.example.demo.model;
-
-public enum Role {
-    ADMIN,
-    EMPLOYEE
-}
