--- src/main/java/com/example/demo/model/VehicleType.java
@@ -1,8 +1,0 @@
-package com.example.demo.model;
-
-public enum VehicleType {
-    SUV,
-    SEDAN,
-    LARGE,
-    COMPACT
-}
