--- src/main/java/com/example/demo/repository/CustomerRepository.java
@@ -1,10 +1,0 @@
-package com.example.demo.repository;
-
-
-import com.example.demo.model.Customer;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
-}
