--- src/main/java/com/example/demo/repository/EmployeeRepository.java
@@ -1,9 +1,0 @@
-package com.example.demo.repository;
-
-import com.example.demo.model.Employee;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface EmployeeRepository extends JpaRepository<Employee, Long> {
-}
