--- src/main/java/com/example/demo/service/impl/CustomerServiceImpl.java
@@ -1,32 +1,0 @@
-package com.example.demo.service.impl;
-
-import com.example.demo.model.Customer;
-import com.example.demo.repository.CustomerRepository;
-import com.example.demo.service.CustomerService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-public class CustomerServiceImpl implements CustomerService {
-
-    @Autowired
-    CustomerRepository customerRepository;
-
-
-    @Override
-    public Customer getCustomer(Long customerId) {
-        return customerRepository.getById(customerId);
-    }
-
-    @Override
-    public List<Customer> getAllCustomers() {
-        return customerRepository.findAll();
-    }
-
-    @Override
-    public void deleteCustomer(Long customerId) {
-        customerRepository.deleteById(customerId);
-    }
-}
