+++ src/main/java/edu/miu/project/configuration/CustomAuthenticationProvider.java
@@ -1,0 +1,44 @@
+package edu.miu.project.configuration;
+
+import edu.miu.project.model.Employee;
+import edu.miu.project.service.EmployeeService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Component;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Component
+public class CustomAuthenticationProvider implements AuthenticationProvider {
+
+    @Autowired
+    private EmployeeService employeeService;
+
+    @Autowired
+    private PasswordEncoder passwordEncoder;
+
+    @Override
+    public Authentication authenticate(Authentication authentication) throws AuthenticationException {
+
+        String userName = authentication.getName();
+        String password = authentication.getCredentials().toString();
+
+        Employee employee = employeeService.findEmployeeByUserName(userName);
+
+        if (employee != null && userName.equals(employee.getUserName()) && passwordEncoder.matches(password, employee.getPassword())) {
+            return new UsernamePasswordAuthenticationToken(userName, password, new ArrayList<>());
+        } else return null;
+    }
+
+    @Override
+    public boolean supports(Class<?> authentication) {
+        return authentication.equals(UsernamePasswordAuthenticationToken.class);
+    }
+}
