+++ src/main/java/edu/miu/project/model/Address.java
@@ -1,0 +1,18 @@
+package edu.miu.project.model;
+
+import lombok.*;
+
+import javax.persistence.Embeddable;
+
+@Embeddable
+@Getter
+@ToString
+@AllArgsConstructor
+@NoArgsConstructor
+@EqualsAndHashCode
+public class Address {
+    private String state;
+    private String city;
+    private String street;
+    private String zipCode;
+}
