+++ src/main/java/edu/miu/project/model/Branch.java
@@ -1,0 +1,7 @@
+package edu.miu.project.model;
+
+public enum Branch {
+    LOS_ANGELES1,
+    LOS_ANGELES2,
+    LOS_ANGELES3
+}
