+++ src/main/java/edu/miu/project/model/FuelType.java
@@ -1,0 +1,7 @@
+package edu.miu.project.model;
+
+public enum FuelType {
+    GASOLINE,
+    DIESEL,
+    ELECTRIC
+}
