+++ src/main/java/edu/miu/project/model/ReservationStatus.java
@@ -1,0 +1,7 @@
+package edu.miu.project.model;
+
+public enum ReservationStatus {
+    RESERVED,
+    PICKED_UP,
+    COMPLETED
+}
