+++ src/main/java/edu/miu/project/model/Role.java
@@ -1,0 +1,6 @@
+package edu.miu.project.model;
+
+public enum Role {
+    ADMIN,
+    EMPLOYEE
+}
