+++ src/main/java/edu/miu/project/model/VehicleType.java
@@ -1,0 +1,8 @@
+package edu.miu.project.model;
+
+public enum VehicleType {
+    SUV,
+    SEDAN,
+    LARGE,
+    COMPACT
+}
