+++ src/main/java/edu/miu/project/repository/CustomerRepository.java
@@ -1,0 +1,10 @@
+package edu.miu.project.repository;
+
+
+import edu.miu.project.model.Customer;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
+}
