+++ src/main/java/edu/miu/project/repository/EmployeeRepository.java
@@ -1,0 +1,10 @@
+package edu.miu.project.repository;
+
+import edu.miu.project.model.Employee;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
+    Employee findEmployeeByUserName(String userName);
+}
