+++ src/main/java/edu/miu/project/service/impl/CustomerServiceImpl.java
@@ -1,0 +1,32 @@
+package edu.miu.project.service.impl;
+
+import edu.miu.project.model.Customer;
+import edu.miu.project.repository.CustomerRepository;
+import edu.miu.project.service.CustomerService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class CustomerServiceImpl implements CustomerService {
+
+    @Autowired
+    CustomerRepository customerRepository;
+
+
+    @Override
+    public Customer getCustomer(Long customerId) {
+        return customerRepository.getById(customerId);
+    }
+
+    @Override
+    public List<Customer> getAllCustomers() {
+        return customerRepository.findAll();
+    }
+
+    @Override
+    public void deleteCustomer(Long customerId) {
+        customerRepository.deleteById(customerId);
+    }
+}
