+++ src/main/resources/application.properties
@@ -11,7 +11,7 @@ server.port=8082
 ## Using MySQL database
 spring.datasource.url = jdbc:mysql://localhost:3306/car_rental_db
 spring.datasource.username = root
-spring.datasource.password = tensaeMySQL
+spring.datasource.password = Welcom@5
 
 
 # ===============================
