--- src/test/java/com/example/demo/Demo1ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.demo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Demo1ApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
