+++ src/test/java/edu/miu/project/Demo1ApplicationTests.java
@@ -1,0 +1,13 @@
+package edu.miu.project;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class Demo1ApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
